+++ mcc2/Assembly/AssemblyType.cs
@@ -5,7 +5,7 @@ public abstract record AssemblyType
     public record Longword() : AssemblyType;
     public record Quadword() : AssemblyType;
     public record Double() : AssemblyType;
-    public record ByteArray(int Size, int Alignment) : AssemblyType;
+    public record ByteArray(long Size, long Alignment) : AssemblyType;
     public record Byte() : AssemblyType;
 
     private AssemblyType() { }
