+++ mcc2/Type.cs
@@ -10,7 +10,7 @@ public abstract record Type
     public record Double() : Type;
     public record FunctionType(List<Type> Parameters, Type Return) : Type;
     public record Pointer(Type Referenced) : Type;
-    public record Array(Type Element, int Size) : Type;
+    public record Array(Type Element, long Size) : Type;
     public record Char() : Type;
     public record SChar() : Type;
     public record UChar() : Type;
